skye link/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-dotenv>=1.0",
  "pydantic>=2.8",
  "httpx>=0.27"
]

[tool.uvicorn]
reload = true